PASCAL VOC style mAP and mAP_BEV

+++ utils/object_detection/losses.py
@@ -97,7 +97,7 @@ class RPNClassificationAndRegressionLoss(nn.Module):
         pos_gt_indices = gt_indices[pos_anchor_indices]
 
         # Balanced sampling
-        num_pos = self.positives_ratio * self.n_cls
+        num_pos = int(self.positives_ratio * self.n_cls)
         num_pos = min(num_pos, pos_anchor_indices.numel())
         num_neg = int(num_pos / self.positives_ratio * self.negatives_ratio)
         num_neg = min(num_neg, neg_anchor_indices.numel())
